+++ src/main/java/gov/vha/isaac/rest/api1/data/mapping/RestMappingSetDisplayFieldBase.java
@@ -18,27 +18,14 @@
  */
 package gov.vha.isaac.rest.api1.data.mapping;
 
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
-
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
-
-import gov.vha.isaac.ochre.api.Get;
-import gov.vha.isaac.ochre.api.LookupService;
-import gov.vha.isaac.ochre.api.component.concept.ConceptChronology;
-import gov.vha.isaac.ochre.api.component.concept.ConceptVersion;
-import gov.vha.isaac.ochre.impl.utility.Frills;
-import gov.vha.isaac.rest.api.exceptions.RestException;
+import gov.vha.isaac.ochre.api.identity.IdentifiedObject;
 import gov.vha.isaac.rest.api1.data.enumerations.MapSetItemComponent;
 import gov.vha.isaac.rest.api1.data.enumerations.RestMapSetItemComponentType;
-import gov.vha.isaac.rest.session.MapSetDisplayFieldsService;
 
 /**
  * 
@@ -60,10 +47,7 @@ public class RestMappingSetDisplayFieldBase
 	 *  1) when componentType below is set to ITEM_EXTENDED, this should be the integer column number that represents the columnPosition 
 	 *     of the extended field (extensionValue.columnNumber)
 	 *  2)  when fieldComponentType below is set to a value such as SOURCE or TARGET - This required value must be an ID pulled from 
-	 *        1/mapping/fields[.id].
-	 *      2a) This id may be a UUID, or
-	 *      2b) This id may be a string constant such as DESCRIPTION
-	 *      
+	 *        1/mapping/fields[.id]. - and will be returned as a UUID.
 	 * the ID returned should be utilized to link the column position of this display field in the RestMappingSetVersion.displayFields list 
 	 * with the ID found in RestMappingItemComputedDisplayField, when placing items on screen.
 	 */
@@ -72,9 +56,12 @@ public class RestMappingSetDisplayFieldBase
 	public String id;
 
 	/**
-	 * A value that describes the type of this display field - which will come from the available values at /1/mapping/fieldComponentTypes 
+	 * An optional value that describes the type of this display field - which will come from the available values at /1/mapping/fieldComponentTypes 
 	 * 
 	 * Example values of for this field are SOURCE, ITEM_EXTENDED, etc
+	 * 
+	 * This is only populated in the context of a mapset that already has a field order assigned.  When this class is returned in a way that it only 
+	 * expresses the possible display fields, this will be null.
 	 */
 	@XmlElement
 	@JsonInclude(JsonInclude.Include.NON_NULL)
@@ -85,60 +72,43 @@ public class RestMappingSetDisplayFieldBase
 		//for Jaxb
 		super();
 	}
-
-	/**
-	 * This is the only constructor that should be called by derived classes
-	 * 
-	 * @param id required to be one of the values returned by MapSetDisplayFieldsService.getAllFieldNames()
-	 * @param componentType required to be non null only for RestMappingSetDisplayFieldCreate
-	 * @throws RestException
-	 */
-	public RestMappingSetDisplayFieldBase(String id, MapSetItemComponent componentType) throws RestException
+	
+	public RestMappingSetDisplayFieldBase(int extendedColumnIdentifier, MapSetItemComponent componentType)
 	{
 		//for Jaxb
 		super();
 
-		if (componentType != null) {
-			// This should only be null when returned as a container for id and description
-			// TODO: Joel fix this by changing hierarchy
-			this.componentType = new RestMapSetItemComponentType(componentType);
+		if (componentType != MapSetItemComponent.ITEM_EXTENDED)
+		{
+			throw new RuntimeException("oops");
 		}
+		this.componentType = new RestMapSetItemComponentType(componentType);
+		if (extendedColumnIdentifier < 0)
+		{
+			throw new RuntimeException("Invalid (negative) map item extended field column " + extendedColumnIdentifier + " for RestMapSetItemComponentType " + componentType);
+		}
+		this.id = extendedColumnIdentifier + "";
 		
-		if (componentType == MapSetItemComponent.ITEM_EXTENDED) {
-			// This only validates the id as a valid non-negative integer,
-			// not that it corresponds to an actual extended field DynamicSememeData element
-			Integer idAsIntegerExtendedFieldsColumnNumber = null;
-			try {
-				idAsIntegerExtendedFieldsColumnNumber = Integer.valueOf(id);
-				if (idAsIntegerExtendedFieldsColumnNumber < 0) {
-					throw new RuntimeException("Invalid (negative) map item extended field column " + id + " for RestMapSetItemComponentType " + componentType);
-				} else {
-					this.id = id;
-				}
-			} catch (NumberFormatException e) {
-				throw new RuntimeException("Invalid map item extended field column " + id + " for RestMapSetItemComponentType " + componentType, e);
-			}
-		} else {
-			// These are globals
-			MapSetDisplayFieldsService service = LookupService.getService(MapSetDisplayFieldsService.class);
-			MapSetDisplayFieldsService.Field field = service.getFieldByConceptIdOrStringIdIfNotConceptId(id);
-			if (field == null) {
-				Optional<? extends ConceptChronology<? extends ConceptVersion<?>>> cc = Frills.getConceptForUnknownIdentifier(id);
-				throw new RestException("RestMappingSetFieldBase.name", id, "Unsupported MapSet field \"" + id + "\"" + (cc.isPresent() ? " (" + Get.conceptDescriptionText(cc.get().getNid()) + ") " : "") + "\". Should be one of " + getFieldNamesWithDescriptions(service.getAllFields()));
-			} else {
-				this.id = field.getId();
-			}
-		}
 	}
 
-	private static Map<Object, String> getFieldNamesWithDescriptions(Collection<MapSetDisplayFieldsService.Field> fields) {
-		Map<Object, String> descriptionsByName = new HashMap<>();
-		
-		for (MapSetDisplayFieldsService.Field field : fields) {
-			descriptionsByName.put(field.getId(), field.getObject() != null ? Get.conceptDescriptionText(field.getObject().getNid()) : MapSetDisplayFieldsService.Field.NonConceptFieldName.valueOf(field.getId()).getDescription());
+	/**
+	 * 
+	 * @param id
+	 * @param componentType should only be null when returning field capabilities - should always be populated in the context of a mapset or mapitem
+	 */
+	public RestMappingSetDisplayFieldBase(IdentifiedObject id, MapSetItemComponent componentType)
+	{
+		super();
+
+		//Null, in cases of communicating capabilities.  should be populated when describing a mapset.
+		if (componentType != null)
+		{
+			this.componentType = new RestMapSetItemComponentType(componentType);
+			if (componentType == MapSetItemComponent.ITEM_EXTENDED) {
+				throw new RuntimeException("oops");
+			}
 		}
-		
-		return descriptionsByName;
+		this.id = id.getPrimordialUuid().toString();
 	}
 
 	/* (non-Javadoc)