+++ src/main/java/gov/vha/isaac/rest/api1/mapping/MappingWriteAPIs.java
@@ -26,7 +26,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-
 import javax.annotation.security.RolesAllowed;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
@@ -37,11 +36,9 @@ import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.SecurityContext;
-
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
 import gov.vha.isaac.MetaData;
 import gov.vha.isaac.ochre.api.Get;
 import gov.vha.isaac.ochre.api.LookupService;
@@ -73,6 +70,7 @@ import gov.vha.isaac.ochre.api.component.sememe.version.dynamicSememe.dataTypes.
 import gov.vha.isaac.ochre.api.constants.DynamicSememeConstants;
 import gov.vha.isaac.ochre.api.coordinate.EditCoordinate;
 import gov.vha.isaac.ochre.api.coordinate.StampCoordinate;
+import gov.vha.isaac.ochre.api.identity.IdentifiedObject;
 import gov.vha.isaac.ochre.api.util.UuidT5Generator;
 import gov.vha.isaac.ochre.impl.utility.Frills;
 import gov.vha.isaac.ochre.mapping.constants.IsaacMappingConstants;
@@ -573,12 +571,12 @@ public class MappingWriteAPIs
 			}
 		} else {
 			MapSetDisplayFieldsService service = LookupService.getService(MapSetDisplayFieldsService.class);
-			MapSetDisplayFieldsService.Field existingField = service.getFieldByConceptIdOrStringIdIfNotConceptId(passedField.id);
+			IdentifiedObject existingField = service.getFieldConceptIdentifierByFieldConceptId(passedField.id);
 			if (existingField == null) {
 				throw new RestException("RestMappingSetFieldCreate.id", passedField.id, "Invalid or unsupported map set field id " + passedField.id 
 						+ " for component type " + msit.name() + ". Must be one of " + service.getAllGlobalFieldIds());
 			}
-			dataString = existingField.getId() + ":" + msit.name();
+			dataString = existingField.getPrimordialUuid().toString() + ":" + msit.name();
 		}
 
 		return new DynamicSememeStringImpl(dataString);
@@ -773,13 +771,18 @@ public class MappingWriteAPIs
 		// If no display fields passed, then add default display fields
 		// as SOURCE, TARGET and EQUIVALENCE_TYPE DESCRIPTION followed by item extended fields, if any
 		if (itemDisplayFields == null || itemDisplayFields.size() == 0) {
+						
 			itemDisplayFields = new ArrayList<>();
-			itemDisplayFields.add(new RestMappingSetDisplayFieldCreate(MapSetDisplayFieldsService.Field.NonConceptFieldName.DESCRIPTION.name(), MapSetItemComponent.SOURCE));
-			itemDisplayFields.add(new RestMappingSetDisplayFieldCreate(MapSetDisplayFieldsService.Field.NonConceptFieldName.DESCRIPTION.name(), MapSetItemComponent.TARGET));
-			itemDisplayFields.add(new RestMappingSetDisplayFieldCreate(MapSetDisplayFieldsService.Field.NonConceptFieldName.DESCRIPTION.name(), MapSetItemComponent.EQUIVALENCE_TYPE));
+			itemDisplayFields.add(new RestMappingSetDisplayFieldCreate(IsaacMappingConstants.get().MAPPING_CODE_DESCRIPTION.getPrimordialUuid().toString(), 
+					MapSetItemComponent.SOURCE));
+			itemDisplayFields.add(new RestMappingSetDisplayFieldCreate(IsaacMappingConstants.get().MAPPING_CODE_DESCRIPTION.getPrimordialUuid().toString(), 
+					MapSetItemComponent.TARGET));
+			itemDisplayFields.add(new RestMappingSetDisplayFieldCreate(IsaacMappingConstants.get().MAPPING_CODE_DESCRIPTION.getPrimordialUuid().toString(), 
+					MapSetItemComponent.EQUIVALENCE_TYPE));
 	
 			for (RestDynamicSememeColumnInfoCreate itemExtendedFieldCol : itemExtendedFieldDefinitions) {
-				itemDisplayFields.add(new RestMappingSetDisplayFieldCreate(calculatedColumnByPassedItemExtendedFieldDefinition.get(itemExtendedFieldCol) + "", MapSetItemComponent.ITEM_EXTENDED));
+				itemDisplayFields.add(new RestMappingSetDisplayFieldCreate(calculatedColumnByPassedItemExtendedFieldDefinition.get(itemExtendedFieldCol) + "", 
+						MapSetItemComponent.ITEM_EXTENDED));
 			}
 		}