+++ ChangeLog.md
@@ -6,7 +6,9 @@ on all changes - including bug fixes.  The API Changelog will only contain docum
 API version number, not the release version number.
 
 * 2017/02/?? - 1.39 - PENDING
-    * 
+    * Internally refatoring the computed display column code to simplify
+    * Changed the way that the RestMappingSetDisplayField.description was calulated, so that now for mapsets it will return useful values for
+        the column headers - by saying things like "source description" or "target description" rather than saying SOURCE with no context.
 
 * 2017/02/03 - 1.38
     * Fixed the root cause of 439352, which was (sometimes) occurring when isaac-rest was deployed - especially if it needed to download a new database.